feat: Implement first-run setup functionality and user creation process
